requirements.txt
functions-framework
firebase-admin
google-cloud-firestore
requests